Hotfix: Get drivers outside of the test folder
