--- src/main/java/com/revature/dartcart/DartCartApplication.java
@@ -1,13 +1,0 @@
-package com.revature.dartcart;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class DartCartApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(DartCartApplication.class, args);
-	}
-
-}
