+++ src/main/java/revature/driver/DartCartApplicationTests.java
@@ -1,4 +1,4 @@
-package com.revature.dartcart;
+package revature.driver;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
