+++ src/main/java/revature/models/User.java
@@ -1,5 +1,5 @@
 
-package com.revature.models;
+package revature.models;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
